requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Caching==2.1.0
requests==2.31.0
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
urllib3==2.0.7
async-timeout==4.0.3
typing-extensions==4.8.0
